requirements.txt
python-telegram-bot==21.7
python-dotenv==1.0.1
google-generativeai==0.8.3
requests==2.32.3
protobuf~=5.28.3